Added ping skeleton code

+++ network/src/main/java/de/haevn/utils/network/PingResult.java
@@ -1,7 +1,7 @@
 package de.haevn.utils.network;
 
-public record PingResult(String host, String hostAddress, long ttl, boolean reachable) {
+public record PingResult(String host, String hostAddress, double ttl, boolean reachable) {
     public String printf(){
-        return String.format("PingResult{host=%s, hostAddress=%s, ttl=%d, reachable=%b}", host, hostAddress, ttl, reachable);
+        return String.format("PingResult{host=%s, hostAddress=%s, ttl=%.0f, reachable=%b}", host, hostAddress, ttl, reachable);
     }
 }
